Moved the colors into another package
