+++ src/main/com/spankingrpgs/crimsonglow/appearance/HairColor.java
@@ -15,7 +15,7 @@
  *      You should have received a copy of the GNU General Public License
  *      along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.spankingrpgs.crimsonglow;
+package com.spankingrpgs.crimsonglow.appearance;
 
 import com.spankingrpgs.model.characters.AppearanceElement;
 import com.spankingrpgs.util.ListUtils;
